Update code for Business of User

+++ Code/company/USR/User/bus/VSoft.Company.USR.User.Business.Dto/Request/UserDeleteDtoRequest.cs
@@ -3,7 +3,7 @@ using VegunSoft.Framework.Business.Dto.Request;
 
 namespace VSoft.Company.USR.User.Business.Dto.Request
 {
-    public class UserDeleteDtoRequest : MDtoRequestDelete<long>
+    public class UserDeleteDtoRequest : MDtoRequestDelete<int>
     {
