+++ Code/company/USR/User/bus/VSoft.Company.USR.User.Business.Dto/Request/UserSaveRangeDtoRequest.cs
@@ -3,7 +3,7 @@ using VSoft.Company.USR.User.Business.Dto.Data;
 
 namespace VSoft.Company.USR.User.Business.Dto.Request
 {
-    public class UserSaveRangeDtoRequest : MDtoRequestRanges<UserDto, long>
+    public class UserSaveRangeDtoRequest : MDtoRequestRanges<UserDto, int>
     {
       
     }
